refactor: DueBadge 공통 컴포넌트 추출 및 ProjectListItem 타입 업데이트
