+++ app/(main)/chat/[projectId]/_components/Layout/ChatProjectSummary.tsx
@@ -1,4 +1,5 @@
 import { ChatProjectSummaryKebab } from "./ChatProjectSummaryKebab";
+import { DueBadge } from "@/app/_components/DueBadge";
 
 interface ChatProjectSummaryProps {
   company: string;
@@ -16,38 +17,6 @@ function formatDate(dateString: string): string {
   return `${year}.${month}.${day}`;
 }
 
-function getDueBadge(dueDate: string | null | undefined): {
-  label: string;
-  className: string;
-} {
-  if (!dueDate) {
-    return {
-      label: "상시",
-      className: "bg-primary-20 text-primary-200",
-    };
-  }
-
-  const today = new Date();
-  today.setHours(0, 0, 0, 0);
-  const due = new Date(dueDate);
-  due.setHours(0, 0, 0, 0);
-  const diff = Math.ceil(
-    (due.getTime() - today.getTime()) / (1000 * 60 * 60 * 24),
-  );
-
-  if (diff < 0) {
-    return {
-      label: "마감",
-      className: "bg-gray-20 text-gray-200",
-    };
-  }
-
-  return {
-    label: diff === 0 ? "D-Day" : `D-${diff}`,
-    className: "bg-primary-20 text-primary-200",
-  };
-}
-
 export function ChatProjectSummary({
   company,
   jobPosition,
@@ -56,7 +25,6 @@ export function ChatProjectSummary({
   onDelete,
 }: ChatProjectSummaryProps) {
   const title = jobPosition ? `${company}_${jobPosition}` : company;
-  const badge = getDueBadge(dueDate);
 
   return (
     <div className="flex items-center justify-between">
@@ -67,11 +35,7 @@ export function ChatProjectSummary({
             {formatDate(dueDate)}
           </span>
         )}
-        <span
-          className={`flex items-center justify-center h-7 min-w-12.5 px-2.5 rounded-lg font-semibold text-base leading-140 ${badge.className}`}
-        >
-          {badge.label}
-        </span>
+        <DueBadge dueDate={dueDate} />
         <ChatProjectSummaryKebab onEdit={onEdit} onDelete={onDelete} />
       </div>
     </div>
